adding more debug issue for innteger arithmetic overflow

+++ NachOS-4.0/code/userprog/ksyscall.h
@@ -22,7 +22,7 @@ void SysHalt()
 }
 
 
-int SysAdd(int op1, int op2)
+long long SysAdd(long long op1, long long op2)
 {
   return op1 + op2;
 }
